+++ docs/changelog.md
@@ -1,12 +1,21 @@
 ## Branches
 Project branch documentation, you will be able to see all the branches created in the project .md:
 ### Create new Branch
+feature, fix, bug
 ```bash 
 git checkout -b nameBranch
 ```
 ### Branches
 - frontend/feature/3148/JuanCardona
 - frontend/feature/3157/ForlanOrdonez
+- frontend/fix/3148/JuanCardona
 -
 -
--
+
+
+### Ejemplo de commit 
+
+fix, feat, doc, styles
+```bash 
+git feat:number_done or WIP
+```